Solves the overflow clipping issue in the secondary stories
This one is interesting. I didn't catch it before I started watching the video. In addition to the solution proposed, I was able to come up with 2 other possible solutions.

1. Add a `position: relative` to the article, and position the paragraph absolutely with no insets.
2. Add—like in this commit—an `align-self: start` to the paragraph, to prevent it stretching to fill the height of the row (which is dictated by the image)

I'm not sure if these approaches have any major downsides compared to the appraoch where you just wrap the paragraph in another div. I like avoiding the extra markup, if possible. I think the `align-self: start` is equivalent in terms of what is ultimately being considered for the height of that <p>.

The absolutely positioned paragraph is fine in the instance where you don't add any more grid children to the given secondary story. But, if you were to add more children, that first additional child would go in the same grid row/column as the original <p> would have occupied if it weren't absolutely positioned, so you'd end up with element overlap.